Make sure plugins are pluged in the same order they were loaded.
Since pieces of hardware are added on a stack, the piece of hardware on
top of it is the last that have been added. Putting it at the beginning
of the hardware array is not a problem, but putting it at the end of it is
more consistent and eases the use of control consoles for plugins that
loads multiple pieces of hardware.